refactor(ui): extract shared SkillEvaluationBadge component
The SkillEvaluationBadge component, its markup, and the evalStatus pass/
borderline/fail thresholds (>=70 / >=50) were copy-pasted into both
KnowledgeNodeDetail.tsx and TacitKnowledgeView.tsx. Extract a single shared
component at src/components/knowledge/SkillEvaluationBadge.tsx (+ co-located
CSS) with the markup and thresholds defined once, parameterizing the CSS class
prefix via a `prefix` prop so each host view keeps its own visual styling.
Both views now import it. Added the missing tk-eval-issues / tk-eval-hints CSS
rules so the unified markup renders correctly in the Tacit Knowledge view.
